New maps added. Transitions between screens initialized.
