+++ deliverables/UML/ObligatoriskOppgave2UML.ucls
@@ -20,7 +20,7 @@
       <operations public="false" package="false" protected="false" private="false" static="false"/>    
     </display>  
   </class>  
-  <class id="3" language="java" name="inf112.RoboRally.app.views.GameScreen" project="INF112"
+  <class id="3" language="java" name="inf112.RoboRally.app.views.Screens.GameScreen" project="INF112"
     file="/INF112/src/main/java/inf112/RoboRally/app/views/showBoard.java" binary="false" corner="BOTTOM_RIGHT">    
     <position height="-1" width="-1" x="834" y="207"/>    
     <display autosize="true" stereotype="true" package="false" initial-value="false" signature="true"
